Update trails schema
Refactor classes
add EndPoint data

+++ TrailFinder.Core/DTOs/Trails/TrailFilterDto.cs
@@ -8,10 +8,10 @@ namespace TrailFinder.Core.DTOs.Trails;
 public record TrailFilterDto(
     string? SearchTerm = null,
     Guid? ParentId = null,
-    decimal? MinDistance = null,
-    decimal? MaxDistance = null,
-    decimal? MinElevation = null,
-    decimal? MaxElevation = null,
+    double? MinDistance = null,
+    double? MaxDistance = null,
+    double? MinElevation = null,
+    double? MaxElevation = null,
     DifficultyLevel? DifficultyLevel = null,
     bool IncludeRouteGeometry = false
 ) : PageRequest;
